requirements.txt
EdgeGPT
httpx[socks]
supervisor
pproxy
python-daemon
pyotp
PyJWT